refactor: Move matches.js under /services folder
